Refactoring: FirebaseRemoteConfig 인스턴스 생성 위치 MainActivity -> FirebaseRemoteConfigRepository로 이
