+++ app/src/main/java/com/eatssu/android/ui/info/InfoActivity.kt
@@ -1,21 +1,19 @@
-package com.eatssu.android.view.infopage
+package com.eatssu.android.ui.info
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import com.eatssu.android.data.enums.Restaurant
 import com.eatssu.android.databinding.ActivityInfoBinding
-import com.eatssu.android.viewmodel.factory.InfoViewModelFactory
+import com.eatssu.android.data.repository.FirebaseRemoteConfigRepository
 
-class InfoActivity @JvmOverloads constructor() : AppCompatActivity() {
+class InfoActivity : AppCompatActivity() {
 
     private lateinit var binding: ActivityInfoBinding
 
     private lateinit var infoViewModel: InfoViewModel
     private lateinit var restaurantType: Restaurant
     private lateinit var firebaseRemoteConfigRepository: FirebaseRemoteConfigRepository
-
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -28,36 +26,41 @@ class InfoActivity @JvmOverloads constructor() : AppCompatActivity() {
         restaurantType = intent.getSerializableExtra("restaurantType") as Restaurant
         binding.tvName.text = restaurantType.displayName
 
-        infoViewModel.infoList.observe(this, Observer {
+        infoViewModel.infoList.observe(this) {
 
-            when(restaurantType){
+            when (restaurantType) {
                 Restaurant.DODAM -> {
                     binding.tvLocation.text = infoViewModel.dodamEtc.value
                     binding.tvTime.text = infoViewModel.dodamTime.value
                     binding.tvEtc.text = infoViewModel.dodamEtc.value
                 }
+
                 Restaurant.HAKSIK -> {
                     binding.tvLocation.text = infoViewModel.dodamLocation.value
                     binding.tvTime.text = infoViewModel.haksikTime.value
                     binding.tvEtc.text = infoViewModel.haksikEtc.value
                 }
+
                 Restaurant.FOOD_COURT -> {
                     binding.tvLocation.text = infoViewModel.foodLocation.value
                     binding.tvTime.text = infoViewModel.foodTime.value
                     binding.tvEtc.text = infoViewModel.foodEtc.value
                 }
+
                 Restaurant.SNACK_CORNER -> {
                     binding.tvLocation.text = infoViewModel.snackLocation.value
                     binding.tvTime.text = infoViewModel.snackTime.value
                     binding.tvEtc.text = infoViewModel.snackEtc.value
                 }
+
                 Restaurant.DORMITORY -> {
                     binding.tvLocation.text = infoViewModel.dodamLocation.value
                     binding.tvTime.text = infoViewModel.dormitoryTime.value
                     binding.tvEtc.text = infoViewModel.dormitoryEtc.value
                 }
+
                 else -> {}
             }
-        })
+        }
     }
 }
